Scene  및 모드 구조 재설계

+++ Objective-D/GameObject.cpp
@@ -200,7 +200,7 @@ void GameObject::ResetAnimationTime(FBXMesh& TargetMesh) {
 
 // FBX 렌더용 함수이다. 앞으로 FBX 모델 렌더링 시 이 함수를 쓰도록 한다.
 // Render3D는 기존의 코드를 최대한 건들지 않기 위해 남겨둔다.
-void GameObject::RenderFBX(FBXMesh& TargetMesh, Texture* TexturePtr, float AlphaValue, float DepthTestFlag){
+void GameObject::RenderFBX(FBXMesh& TargetMesh, Texture* TexturePtr, float AlphaValue, int DepthTestFlag){
 	for (auto const& M : TargetMesh.MeshPart)
 		Render3D(M, TexturePtr, AlphaValue, DepthTestFlag);
 }
